requirements.txt
torch
torchvision
timm
flask
opencv-python
numpy
pillow
matplotlib
scikit-learn
grad-cam
onnx
onnxruntime